Add files via upload
It is a database project based on a university management system, which has twelve entities and each entity has several attributes with one of them being a Primary Key and each attribute has some kind of constraint applied on it as well as some foreign keys. We have tried our level best to create an effective management system.